requirements.txt
PyQt6-Qt6==6.7.0
python-dotenv==1.0.1
aiofiles==23.2.1
yadisk==2.1.0
httpx==0.27.0
apscheduler==3.10.4